requirements.txt
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
google-auth==2.15.0
google-auth-oauthlib==0.8.0
idna==3.4
itsdangerous==2.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
